# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Omnibase REST API
  version: 0.9.18
  description: |
    Self-hostable Backend-as-a-Service providing database management, authentication, payments, storage, and email services.

    ## Features
    - **Database**: PostgreSQL with RLS and migrations
    - **Authentication**: Ory Kratos integration with session management
    - **Payments**: Stripe integration with version-controlled billing configs
    - **Storage**: S3-compatible object storage with RLS
    - **Email**: Transactional email service
    - **Permissions**: Fine-grained access control via Ory Keto

    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Use the appropriate security scheme based on the endpoint requirements.
  contact:
    name: Omnibase Support
    url: https://omnibase.dev/support
    email: support@omnibase.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://omnibase.dev/terms
servers:
  - url: https://api.omnibase.tech
    description: Production server
...
paths:
...
  /api/v1/permissions/check:
    post:
      operationId: checkPermission
      summary: Check permission
      description: |
        Checks if a subject has a specific permission on an object using Ory Keto.

        ## Authentication
        Requires session authentication.

        ## Request Format
        Provide either `subject_id` or `subject_set` (not both). The request uses oneOf 
        to enforce this constraint at the schema level.

        ## Use Cases
        - Verify user permissions before performing actions
        - Implement fine-grained access control
        - Check role-based permissions
      tags:
        - V1 Permissions
      security:
        - CookieAuth: []
          SessionTokenAuth: []
          ServiceKeyAuth: []
      requestBody:
        required: true
        description: Permission check request with either subject_id or subject_set
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckPermissionRequest'
            examples:
              withSubjectId:
                summary: Check permission with subject_id
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_invite_user
                  subject_id: 550e8400-e29b-41d4-a716-446655440000
              withSubjectSet:
                summary: Check permission with subject_set
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_delete_tenant
                  subject_set:
                    namespace: Tenant
                    object: tenant_parent_456
                    relation: can_invite_user
      responses:
        '200':
          description: Permission check result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CheckPermissionResponse'
              examples:
                allowed:
                  summary: Permission granted
                  value:
                    status: 200
                    data:
                      allowed: true
                denied:
                  summary: Permission denied
                  value:
                    status: 200
                    data:
                      allowed: false
        '400':
          description: Invalid request body - namespace, object, and relation cannot be empty, or both/neither subject fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
              examples:
                bothSubjects:
                  summary: Both subject_id and subject_set provided
                  value:
                    status: 400
                    error: Provide either subject_id or subject_set, not both
                neitherSubject:
                  summary: Neither subject_id nor subject_set provided
                  value:
                    status: 400
                    error: Either subject_id or subject_set must be provided
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Failed to check permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
...
  /api/v1/tenants/roles:
    get:
      tags:
        - V1 Tenants
      summary: List roles
      description: |
        Returns all roles for the authenticated tenant, including both system roles and custom tenant-specific roles.

        ## Authentication
        - **Session Auth**: Requires JWT token / Cookie Session with tenant context
        - **Service Key Auth**: Requires X-Service-Key + X-Tenant-ID header

        ## Use Cases
        - Display available roles to assign
        - Role management UI
        - Permission auditing
      operationId: listRoles
      parameters:
        - name: X-Tenant-Id
          in: header
          required: false
          description: Tenant ID (UUID) - Required when using X-Service-Key header
          schema:
            type: string
            format: uuid
      security:
        - CookieAuth: []
        - SessionTokenAuth: []
        - ServiceKeyAuth: []
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RolesListResponse'
        '400':
          description: Missing tenant ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to list roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - V1 Tenants
      summary: Create role
      description: |
        Creates a new custom role for the tenant with specified permissions.

        ## Authentication
        - **Session Auth**: Requires JWT token / Cookie Session with tenant context and appropriate permissions
        - **Service Key Auth**: Requires X-Service-Key + X-Tenant-ID + X-User-ID headers with appropriate permissions

        ## Permission Format
        Permissions should be in the format: `namespace:resource#relation`
        - Tenant-wide: `tenant#relation`
        - Resource-specific: `project:uuid#relation`

        ## Use Cases
        - Create custom roles for specific workflows
        - Define project-specific permissions
        - Build granular access control
      operationId: createRole
      parameters:
        - name: X-User-Id
          in: header
          required: false
          description: User ID (UUID) - Required when using X-Service-Key header
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        - name: X-Tenant-Id
          in: header
          required: false
          description: Tenant ID (UUID) - Required when using X-Service-Key header
          schema:
            type: string
            format: uuid
          example: 31c60057-bd7c-41b8-b96e-c4ceb845034f
      security:
        - CookieAuth: []
        - SessionTokenAuth: []
        - ServiceKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '200':
          description: Role created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          description: Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Role with this name already exists for tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/schemas-ConflictResponse'
              example:
                status: 409
                error: Role with this name already exists for tenant
        '500':
          description: Failed to create role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: |
        Session cookie authentication. Cookie name: `ory_kratos_session`.
        Automatically set by browser after Kratos login.
    SessionTokenAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Kratos session JWT token. Alternative to cookie authentication for
        non-browser clients. Obtain from Kratos after login.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-Service-Key
      description: |
        Service-to-service authentication key for backend operations. When
        used with tenant endpoints, must include X-Tenant-ID header.
  schemas:
    SuccessResponse:
      type: object
      description: Standard success response wrapper
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 200
        data:
          description: Response data payload
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: boolean
      required:
        - status
...
    BadRequestResponse:
      type: object
      description: Bad Request error response (400)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message or details
          example: Invalid request parameters
      required:
        - status
        - error
    UnauthorizedResponse:
      type: object
      description: Unauthorized error response (401)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 401
        error:
          type: string
          description: Error message
          example: Unauthorized
      required:
        - status
        - error
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error response (500)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 500
        error:
          type: string
          description: Error message
          example: Internal Server Error
      required:
        - status
        - error
...
    ErrorResponse:
      type: object
      description: Generic error response
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message or details
          example: An error occurred
      required:
        - status
        - error
...
    CheckPermissionRequest:
      description: Check permission request. Must provide exactly one of subject_id or subject_set.
      oneOf:
        - $ref: '#/components/schemas/CheckPermissionRequestWithSubjectId'
        - $ref: '#/components/schemas/CheckPermissionRequestWithSubjectSet'
    CheckPermissionResponse:
      type: object
      description: Permission check result
      properties:
        allowed:
          type: boolean
          description: Whether the permission check passed
          example: true
      required:
        - allowed
...
    ForbiddenResponse:
      type: object
      description: Forbidden error response (403)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 403
        error:
          type: string
          description: Error message
          example: Forbidden
      required:
        - status
        - error
...
    Role:
      type: object
      description: Role definition with permissions
      required:
        - id
        - role_name
        - permissions
        - user_ids
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique role identifier
          example: 7bde7bd1-9be9-42f5-bc4b-be9f24cde432
        tenant_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the tenant this role belongs to (NULL for system roles)
        role_name:
          type: string
          description: Role name
          example: member
        permissions:
          type: array
          description: List of permissions in namespace:resource#relation format
          items:
            type: string
          example:
            - tenant#invite_user
            - tenant#remove_user
        user_ids:
          type: array
          description: Array of user IDs assigned to this role
          items:
            type: string
            format: uuid
          example:
            - 550e8400-e29b-41d4-a716-446655440000
            - 550e8400-e29b-41d4-a716-446655440001
        created_at:
          type: string
          format: date-time
          description: Timestamp when role was created
          example: '2025-11-10T00:18:19.653645Z'
        updated_at:
          type: string
          format: date-time
          description: Timestamp when role was last updated
          example: '2025-11-10T00:33:08.726632Z'
    RolesListResponse:
      type: object
      description: List of roles in a tenant
      required:
        - roles
      properties:
        roles:
          type: array
          description: List of roles (including system roles)
          items:
            $ref: '#/components/schemas/Role'
    CreateRoleRequest:
      type: object
      description: Request to create a new role
      required:
        - role_name
        - permissions
      properties:
        role_name:
          type: string
          minLength: 1
          description: Name of the role (required, cannot be empty)
          example: test_project_viewer
        permissions:
          type: array
          minItems: 1
          description: List of permissions in namespace:resource#relation format (required, must have at least one non-empty permission). Empty strings are not allowed and will be rejected with a 400 error.
          items:
            type: string
            minLength: 1
          example:
            - tenant#read
            - tenant#manage_projects
...
  responses:
    BadRequestError:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestResponse'
        text/plain:
          schema:
            type: string
            example: 400 Bad Request
...
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForbiddenResponse'